+++ clang-tools-extra/clang-tidy/readability/ReadabilityTidyModule.cpp
@@ -44,6 +44,7 @@
 #include "StringCompareCheck.h"
 #include "UniqueptrDeleteReleaseCheck.h"
 #include "UppercaseLiteralSuffixCheck.h"
+#include "VisibleExceptionPropagationCheck.h"
 
 namespace clang {
 namespace tidy {
@@ -122,6 +123,8 @@ public:
         "readability-uniqueptr-delete-release");
     CheckFactories.registerCheck<UppercaseLiteralSuffixCheck>(
         "readability-uppercase-literal-suffix");
+    CheckFactories.registerCheck<VisibleExceptionPropagationCheck>(
+        "readability-visible-exception-propagation");
   }
 };