+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -196,6 +196,12 @@ Improvements to include-fixer
 
 The improvements are...
 
+- New :doc:`readability-visible-exception-propagation
+  <clang-tidy/checks/readability-visible-exception-propagation>` check.
+
+  Issues warnings for call sites of throwing functions that are not annotaed [[try]].
+  FIXME: TYDO
+
 Improvements to clang-include-fixer
 -----------------------------------